# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SHED Motor Loco TM Temperature Monitoring API",
    "version": "1.0.0",
    "description": "HIMNISH LIMITED — Real-time traction-motor temperature monitoring for Indian Railways SHED motor locos. Ingestion is key-authenticated; the dashboard API uses a JWT bearer token from /login."
  },
  "servers": [{ "url": "/api/v1" }],
...
  "paths": {
...
    "/ingest": { "post": { "tags": ["Ingestion"], "summary": "Push readings for a loco (X-API-Key)", "description": "Header X-API-Key required. Optional 'ts' allows replaying buffered readings with their original timestamp.", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "loco_id": { "type": "string" }, "shed_id": { "type": "string" }, "ts": { "type": "string", "format": "date-time" }, "readings": { "type": "array", "items": { "type": "object", "properties": { "sensor_id": { "type": "string" }, "tm_id": { "type": "string" }, "temperature": { "type": "number" }, "battery_health": { "type": "number" }, "signal_strength": { "type": "number" } } } } } } } } }, "responses": { "200": { "description": "accepted" }, "401": { "description": "invalid API key" } } } },
    "/device-config": { "get": { "tags": ["Devices"], "summary": "A field RUT pulls its own config (self-update)", "description": "Auth with the shared bootstrap key.", "parameters": [{ "name": "device", "in": "query", "required": true, "schema": { "type": "string" } }, { "name": "key", "in": "query", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "config for the device" }, "401": { "description": "invalid bootstrap key" }, "404": { "description": "device not registered" } } } },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "apiKey": { "type": "apiKey", "in": "header", "name": "X-API-Key" }
    }
  }
}